Update role, readme and user to fit flyway

+++ bucket/src/main/java/com/expoai/bucket/enums/RoleEnum.java
@@ -9,9 +9,9 @@ public enum RoleEnum {
     SCRAPPER(3),
     ;
 
-    long id ;
+    int id ;
 
-    RoleEnum(long i) {
+    RoleEnum(int i) {
         id = i;
     }
